moved alert menu out of alert rule folder #1055
