+++ Source/Alerts/Web/src/components/Alerts/AlertOverview.js
@@ -8,7 +8,7 @@ import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 import Paper from '@material-ui/core/Paper';
 import Button from '@material-ui/core/Button';
-import MenuAlert from '../AlertsRule/menuAlert';
+import MenuAlert from '../menuAlert';
 import CBSNavigation from '../Navigation/CBSNavigation';
 
 const CustomTableCell = withStyles(theme => ({
